safari-600.7-branch/LayoutTests/mathml/presentation/stretchy-depth-height-symmetric-expected.txt

This test passes if the brace stretches symmetrically and is large enough to cover its sibling.

 
 
 

large depth: PASS
large depth (explicit symmetric attribute): PASS
large height: PASS
large height (explicit symmetric attribute): PASS
